requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-storage==1.35.0